+++ hedera-node/hedera-token-service-impl/src/test/java/com/hedera/node/app/service/token/impl/test/handlers/CryptoApproveAllowanceHandlerTest.java
@@ -20,6 +20,7 @@ import static com.hedera.hapi.node.base.ResponseCodeEnum.EMPTY_ALLOWANCES;
 import static com.hedera.hapi.node.base.ResponseCodeEnum.INVALID_ALLOWANCE_OWNER_ID;
 import static com.hedera.hapi.node.base.ResponseCodeEnum.INVALID_DELEGATING_SPENDER;
 import static com.hedera.hapi.node.base.ResponseCodeEnum.MAX_ALLOWANCES_EXCEEDED;
+import static com.hedera.hapi.node.base.ResponseCodeEnum.NEGATIVE_ALLOWANCE_AMOUNT;
 import static com.hedera.hapi.node.base.ResponseCodeEnum.OK;
 import static com.hedera.hapi.node.base.ResponseCodeEnum.SENDER_DOES_NOT_OWN_NFT_SERIAL_NO;
 import static com.hedera.node.app.spi.fixtures.Assertions.assertThrowsPreCheck;
@@ -420,6 +421,40 @@ class CryptoApproveAllowanceHandlerTest extends CryptoTokenHandlerTestBase {
         assertThat(modifiedOwner.cryptoAllowances()).isEmpty();
         assertThat(modifiedOwner.tokenAllowances()).isEmpty();
         assertThat(existingOwner.approveForAllNftAllowances()).isEmpty();
+    }
+
+    @Test
+    void validateNegativeAmounts() {
+        final var txn = cryptoApproveAllowanceTransaction(
+                payerId,
+                false,
+                List.of(CryptoAllowance.newBuilder()
+                        .amount(-1L)
+                        .owner(ownerId)
+                        .spender(spenderId)
+                        .build()),
+                List.of(tokenAllowance),
+                List.of(nftAllowance));
+
+        assertThatThrownBy(() -> subject.pureChecks(txn))
+                .isInstanceOf(PreCheckException.class)
+                .has(responseCode(NEGATIVE_ALLOWANCE_AMOUNT));
+
+        cryptoApproveAllowanceTransaction(
+                payerId,
+                false,
+                List.of(cryptoAllowance),
+                List.of(TokenAllowance.newBuilder()
+                        .amount(-1L)
+                        .owner(ownerId)
+                        .spender(spenderId)
+                        .tokenId(fungibleTokenId)
+                        .build()),
+                List.of(nftAllowance));
+
+        assertThatThrownBy(() -> subject.pureChecks(txn))
+                .isInstanceOf(PreCheckException.class)
+                .has(responseCode(NEGATIVE_ALLOWANCE_AMOUNT));
     }
 
     @Test
